backend/requirements.txt
fastapi
uvicorn
pyserial
numpy
PyWavelets
python-dotenv
matplotlib
websockets
scipy